fix: add .gitignore config to prevent accidental commital of secrets, etc
Add go.allowlist.gitignore file for gitignore configuration to ensure Git does not commit unwanted files. This file takes a different approach from a typical .gitignore, which instructs Git what files to avoid committing. Instead, this file creates an allowlist, explicitly naming the files and directories that should be committed.

This ensures that people using different OS's, IDEs, etc. which produce different types of extraneous files don't have to constantly add to the .gitignore - unless they are adding an atypical file (one not ending in the conventional extension of `.go`) for the language/framework.